moved virtual-key-refactor to its own branch
