+++ ui/litellm-dashboard/src/app/(dashboard)/components/modals/CreateUserModal.tsx
@@ -29,8 +29,7 @@ import {
 } from "@/components/networking";
 import { getModelDisplayName } from "@/components/key_team_helpers/fetch_available_models_team_key";
 import BulkCreateUsersButton from "@/components/bulk_create_users_button";
-import { fetchTeams } from "@/app/(dashboard)/virtual-keys/networking";
-import useTeams from "@/app/(dashboard)/virtual-keys/hooks/useTeams";
+import useTeams from "@/app/(dashboard)/hooks/useTeams";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
 // Helper function to generate UUID compatible across all environments
