+++ ui/litellm-dashboard/src/app/(dashboard)/hooks/useTeams.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
-import { fetchTeams } from "@/app/(dashboard)/virtual-keys/networking";
 import { Team } from "@/components/key_team_helpers/key_list";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
+import { fetchTeams } from "@/app/(dashboard)/networking";
 
 const useTeams = () => {
   const [teams, setTeams] = useState<Team[] | null>([]);
