--- ui/litellm-dashboard/src/app/(dashboard)/virtual-keys/components/CreateKeyModal/types.ts
@@ -1,17 +1,0 @@
-export interface User {
-  user_id: string;
-  user_email: string;
-  role?: string;
-}
-
-export interface UserOption {
-  label: string;
-  value: string;
-  user: User;
-}
-
-export type Model = { id: string };
-
-export type ModelAvailableResponse = { data: Model[] };
-
-export type ModelAliases = { [key: string]: string };
