+++ ui/litellm-dashboard/src/app/(dashboard)/virtual-keys/page.tsx
@@ -1,16 +1,18 @@
 "use client";
 
-import VirtualKeysTable from "@/app/(dashboard)/virtual-keys/components/VirtualKeysTable/VirtualKeysTable";
 import { useState } from "react";
-import useKeyList from "@/components/key_team_helpers/key_list";
+import useKeyList, { KeyResponse } from "@/components/key_team_helpers/key_list";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { Col, Grid } from "@tremor/react";
-import CreateKey from "@/app/(dashboard)/virtual-keys/components/CreateKey";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
+import UserDashboard from "@/components/user_dashboard";
+import useTeams from "@/app/(dashboard)/hooks/useTeams";
+import { Organization } from "@/components/networking";
 
 const VirtualKeysPage = () => {
-  const { accessToken, userRole } = useAuthorized();
+  const { accessToken, userRole, userId, premiumUser, userEmail } = useAuthorized();
+  const { teams, setTeams } = useTeams();
   const [createClicked, setCreateClicked] = useState<boolean>(false);
+  const [organizations, setOrganizations] = useState<Organization[]>([]);
 
   const queryClient = new QueryClient();
 
@@ -28,20 +30,21 @@ const VirtualKeysPage = () => {
 
   return (
     <QueryClientProvider client={queryClient}>
-      <div className="w-full mx-4 h-[75vh]">
-        <Grid numItems={1} className="gap-2 p-8 w-full mt-2">
-          <Col numColSpan={1} className="flex flex-col gap-2">
-            <CreateKey team={null} userRole={userRole} data={null} addKey={addKey} />
-            <VirtualKeysTable
-              keys={keys}
-              setKeys={setKeys}
-              pagination={pagination}
-              onPageChange={() => {}}
-              organizations={null}
-            />
-          </Col>
-        </Grid>
-      </div>
+      <UserDashboard
+        userID={userId}
+        userRole={userRole}
+        userEmail={userEmail}
+        teams={teams}
+        keys={keys}
+        setUserRole={() => {}}
+        setUserEmail={() => {}}
+        setTeams={setTeams}
+        setKeys={setKeys}
+        premiumUser={premiumUser}
+        organizations={organizations}
+        addKey={addKey}
+        createClicked={createClicked}
+      />
     </QueryClientProvider>
   );
 };
