+++ ui/litellm-dashboard/src/components/view_users.tsx
@@ -30,6 +30,8 @@ import { updateExistingKeys } from "@/utils/dataUtils";
 import { useDebouncedState } from "@tanstack/react-pacer/debouncer";
 import { isAdminRole } from "@/utils/roles";
 import NotificationsManager from "./molecules/notifications_manager";
+import { Setter } from "@/types";
+import { KeyResponse } from "@/components/key_team_helpers/key_list";
 
 interface ViewUserDashboardProps {
   accessToken: string | null;
